Change Loop so it needs to be executed multiple times...
until the limit is reached, then completed will run.

This allows a user to put any task has "delay" behaviour in during the
repeatitions. All they have to do is connect body route back to the start
of loop task.

Another "Repeat" task will be introduced to simulate the traditional "For
Loop"