Fix Getter Setter input output naming issue.
Originally getter and setter node uses variable id as the input and output
name(or you can call it id).

The problem is that variable id will change during actor importing.
Therefore links for getter and setter blocks no longer valid.

The easiest way would be hard code the output and input name to be
"output" and "input". As long as we can ensure it is unique in the
corresponding input list and output list scope, which is obviously
correct.