BIG FIX on mobile device scrolling panning issue!!!!
I don't fully understand the logic behind it, but I've tested on iPad and my Android
mobile. It works!!!!!

But my guess is they are related to the size of the scrollable content.
You have to make the container is 100% fit the screen and make offscreen
part hidden.

Regarding the svg connection line offset issue while dragging:

It is because while dragging, it triggers browser default elastic bouncing
behaviours. Therefore, I just need to preventDefault actions on anything
that are not part of browser(as we still need to scroll browser content).

I chose to disable both BrainGraph.container and
BrainGraph.blockContainer. I try to only preventDefault on
BrainGraph.container, it does not work properly.

It is strange, as container is parent of the blockContainer. I'll try it
again.